+++ app/Http/Controllers/Api/ApiContactController.php
@@ -6,7 +6,6 @@ use Illuminate\Http\Request;
 use App\Helpers\SearchHelper;
 use App\Models\Contact\Contact;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Collection;
 use App\Jobs\UpdateLastConsultedDate;
 use Illuminate\Database\QueryException;
 use App\Services\Contact\Contact\SetMeContact;
@@ -21,7 +20,6 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Http\Resources\Contact\Contact as ContactResource;
 use App\Services\Contact\Contact\UpdateContactIntroduction;
 use App\Services\Contact\Contact\UpdateContactFoodPreferences;
-use App\Http\Resources\Contact\ContactWithContactFields as ContactWithContactFieldsResource;
 
 class ApiContactController extends ApiController
 {
@@ -59,9 +57,7 @@ class ApiContactController extends ApiController
                 return $this->respondInvalidQuery();
             }
 
-            $collection = $this->applyWithParameter($contacts, $this->getWithParameter());
-
-            return $collection->additional([
+            return ContactResource::collection($contacts)->additional([
                 'meta' => [
                     'query' => $needle,
                 ],
@@ -79,7 +75,7 @@ class ApiContactController extends ApiController
             return $this->respondInvalidQuery();
         }
 
-        return $this->applyWithParameter($contacts, $this->getWithParameter());
+        return ContactResource::collection($contacts);
     }
 
     /**
@@ -87,7 +83,7 @@ class ApiContactController extends ApiController
      *
      * @param Request $request
      * @param int $id
-     * @return ContactResource|JsonResponse|ContactWithContactFieldsResource
+     * @return ContactResource|JsonResponse
      */
     public function show(Request $request, int $id)
     {
@@ -100,10 +96,6 @@ class ApiContactController extends ApiController
         }
 
         UpdateLastConsultedDate::dispatch($contact);
-
-        if ($this->getWithParameter() == 'contactfields') {
-            return new ContactWithContactFieldsResource($contact);
-        }
 
         return new ContactResource($contact);
     }
@@ -182,20 +174,6 @@ class ApiContactController extends ApiController
         app(DestroyContact::class)->execute($data);
 
         return $this->respondObjectDeleted($contactId);
-    }
-
-    /**
-     * Apply the `?with=` parameter.
-     * @param  Collection $contacts
-     * @return JsonResource
-     */
-    private function applyWithParameter($contacts, string $parameter = null)
-    {
-        if ($parameter == 'contactfields') {
-            return ContactWithContactFieldsResource::collection($contacts);
-        }
-
-        return ContactResource::collection($contacts);
     }
 
     /**