+++ app/Http/Controllers/DAV/Backend/CalDAV/CalDAVBackend.php
@@ -143,6 +143,8 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
         if ($backend) {
             return $backend->getChanges($calendarId, $syncToken);
         }
+
+        return [];
     }
 
     /**
@@ -191,6 +193,8 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
                 })
                 ->toArray();
         }
+
+        return [];
     }
 
     /**
@@ -219,6 +223,8 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
                 return $backend->prepareData($obj);
             }
         }
+
+        return [];
     }
 
     /**
@@ -262,12 +268,13 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
      * @param string $calendarData
      * @return string|null
      */
-    public function updateCalendarObject($calendarId, $objectUri, $calendarData)
+    public function updateCalendarObject($calendarId, $objectUri, $calendarData): ?string
     {
         $backend = $this->getBackend($calendarId);
-        if ($backend) {
-            return $backend->updateOrCreateCalendarObject($objectUri, $calendarData);
-        }
+
+        return $backend ?
+            $backend->updateOrCreateCalendarObject($objectUri, $calendarData)
+            : null;
     }
 
     /**