+++ app/Http/Controllers/DAV/Backend/CalDAV/ICalDAVBackend.php
@@ -103,7 +103,7 @@ interface ICalDAVBackend
      * @param string $calendarData
      * @return string|null
      */
-    public function updateOrCreateCalendarObject($objectUri, $calendarData);
+    public function updateOrCreateCalendarObject($objectUri, $calendarData): ?string;
 
     /**
      * Deletes an existing calendar object.