+++ app/Http/Controllers/DAV/Backend/CardDAV/CardDAVBackend.php
@@ -198,9 +198,9 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      * Prepare datas for this contact.
      *
      * @param Contact $contact
-     * @return array|null
+     * @return array
      */
-    private function prepareCard($contact)
+    private function prepareCard($contact): array
     {
         try {
             $vcard = app(ExportVCard::class)
@@ -401,6 +401,8 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
             Log::debug(__CLASS__.' updateCard: '.(string) $e);
             throw $e;
         }
+
+        return;
     }
 
     /**
@@ -431,7 +433,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      * @param \Sabre\DAV\PropPatch $propPatch
      * @return bool|null
      */
-    public function updateAddressBook($addressBookId, DAV\PropPatch $propPatch)
+    public function updateAddressBook($addressBookId, DAV\PropPatch $propPatch): ?bool
     {
         $propPatch->handle('{'.CalDAVPlugin::NS_CALENDARSERVER.'}me-card', function ($props) use ($addressBookId) {
             $contact = $this->getObject($addressBookId, $props->getHref());
@@ -446,6 +448,8 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
 
             return true;
         });
+
+        return null;
     }
 
     /**