+++ app/Http/Controllers/DAV/DAVACL/PrincipalBackend.php
@@ -88,6 +88,8 @@ class PrincipalBackend extends AbstractBackend
                 return $principal;
             }
         }
+
+        return [];
     }
 
     /**