+++ app/Http/Resources/Conversation/Message.php
@@ -3,10 +3,13 @@
 namespace App\Http\Resources\Conversation;
 
 use App\Helpers\DateHelper;
-use Illuminate\Http\Resources\Json\Resource;
+use Illuminate\Http\Resources\Json\JsonResource;
 use App\Http\Resources\Contact\ContactShort as ContactShortResource;
 
-class Message extends Resource
+/**
+ * @extends JsonResource<\App\Models\Contact\Message>
+ */
+class Message extends JsonResource
 {
     /**
      * Transform the resource into an array.