+++ app/Http/Resources/Emotion/Emotion.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Resources\Emotion;
 
-use Illuminate\Http\Resources\Json\Resource;
+use Illuminate\Http\Resources\Json\JsonResource;
 
-class Emotion extends Resource
+/**
+ * @extends JsonResource<\App\Models\Instance\Emotion\Emotion>
+ */
+class Emotion extends JsonResource
 {
     /**
      * Transform the resource into an array.